Updated algorithm for better timing.
The TIMING option now spits out a space separated
file with which gnuplot can deal with way better.
A gnuplot script with a simple setup was also added.

Fixed a bug with the calculation of the total
fitness for the walshls solver, the sum of the
S vector entries were taken while the sum of
the Wprime vector entries need to be taken.

The printing during normal operation with SHOW_STEPS
on is now the same between the naive and walshls solver.
Because of this can you now check the correcteness of
the walshls solver against the naive solver by setting
the alpha of the walshls high and executing compare.sh.

A new test problem instance, deceptive trap, was added
which was made for debugging.