Fixed bug in walshls solver.
A bug I found with a case with plateau's is fixed
by using epsilon compare to determine if a S value
is improving. A case had a S value of -2e-16 or something
which caused it to be negative while it is probably
a rounding error.

I also added a lot of comments and made the code more
consistent.